IGNITE-20939 Extract distribution zone integration tests to related module.
