+++ modules/distribution-zones/build.gradle
@@ -51,7 +51,6 @@ dependencies {
     testImplementation project(':ignite-network-api')
     testImplementation project(':ignite-metastorage-api')
     testImplementation project(':ignite-metastorage')
-    testImplementation project(':ignite-runner')
     testImplementation project(':ignite-system-view-api')
 
     testImplementation(testFixtures(project(':ignite-core')))
@@ -76,6 +75,27 @@ dependencies {
     testFixturesImplementation project(':ignite-catalog')
     testFixturesImplementation(testFixtures(project(':ignite-core')))
     testFixturesImplementation(testFixtures(project(':ignite-metastorage')))
+
+    integrationTestImplementation libs.awaitility
+    integrationTestImplementation libs.hamcrest.core
+    integrationTestImplementation libs.jetbrains.annotations
+    integrationTestImplementation project(':ignite-api')
+    integrationTestImplementation project(':ignite-configuration')
+    integrationTestImplementation project(':ignite-metastorage-api')
+    integrationTestImplementation project(':ignite-vault')
+    integrationTestImplementation project(':ignite-table')
+    integrationTestImplementation project(':ignite-catalog')
+    integrationTestImplementation project(':ignite-affinity')
+    integrationTestImplementation project(':ignite-network-api')
+    integrationTestImplementation project(':ignite-network')
+    integrationTestImplementation project(':ignite-system-view-api')
+    integrationTestImplementation testFixtures(project(':ignite-core'))
+    integrationTestImplementation testFixtures(project(':ignite-runner'))
+    integrationTestImplementation testFixtures(project(':ignite-metastorage'))
+    integrationTestImplementation testFixtures(project(':ignite-cluster-management'))
+    integrationTestImplementation testFixtures(project(':ignite-configuration'))
+    integrationTestImplementation testFixtures(project(':ignite-network'))
+    integrationTestImplementation testFixtures(project(':ignite-distribution-zones'))
 }
 
 description = 'ignite-distribution-zones'
