+++ modules/runner/build.gradle
@@ -117,7 +117,6 @@ dependencies {
 
     integrationTestAnnotationProcessor project(':ignite-configuration-annotation-processor')
     integrationTestAnnotationProcessor libs.jmh.annotation.processor
-    integrationTestImplementation project(':ignite-distribution-zones')
     integrationTestImplementation project(':ignite-affinity')
     integrationTestImplementation project(':ignite-jdbc')
     integrationTestImplementation project(':ignite-page-memory')
@@ -194,6 +193,9 @@ dependencies {
     testFixturesImplementation project(':ignite-system-view-api')
     testFixturesImplementation project(':ignite-system-view')
     testFixturesImplementation project(':ignite-storage-api')
+    testFixturesImplementation project(':ignite-cluster-management')
+    testFixturesImplementation project(':ignite-configuration')
+    testFixturesImplementation project(':ignite-vault')
     testFixturesImplementation testFixtures(project(':ignite-api'))
     testFixturesImplementation testFixtures(project(':ignite-core'))
     testFixturesImplementation testFixtures(project(':ignite-sql-engine'))
